Title: 시험 통과 여부 확인하기 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240103/시험 통과 여부 확인하기 2/verify-test-passed-2.java
@@ -15,7 +15,7 @@ public class Main {
                 arr[j] = sc.nextInt();
                 sum+= arr[i];
             }
-            int avg = sum/4;
+            double avg = (double)sum/4;
             if(avg>=60) {
                 System.out.println("pass");
                 person++;
